implement some helper functions
